requirements.txt
tensorflow==2.15.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
aiofiles==23.2.1
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.23
sqlite3
pytest==7.4.3
pytest-asyncio==0.21.1
requests==2.31.0
httpx==0.25.2